pyproject.toml
[project]
name = "askdagger_cliport"
version = "0.1.0"
description = "Code for ASkDAgger CLIPort experiments"
readme = "README.md"
requires-python = "<3.11,>=3.8"
dependencies = [
    "clip @ git+https://github.com/openai/clip.git@main",
    "openai<1.0.0,>=0.22.1",
    "ftfy<7.0.0,>=6.1.1",
    "regex<2023.0.0,>=2022.8.17",
    "imageio==2.9.0",
    "imageio-ffmpeg==0.4.5",
    "meshcat==0.0.18",
    "kornia==0.4.1",
    "opencv-python<5.0.0,>=4.3.0",
    "pybullet<4.0.0,>=3.2.1",
    "gymnasium<1.0.0,>=0.28.1",
    "pytorch-lightning==1.0.3",
    "hydra-core==1.0.5",
    "wandb==0.10.15",
    "transforms3d<1.0.0,>=0.4.1",
    "matplotlib<4.0.0,>=3.6.3",
    "protobuf<4",
    "scikit-learn<2.0.0,>=1.3.1",
    "tqdm<5.0.0,>=4.66.1",
    "torch==1.11.0+cu113",
    "torchvision==0.12.0+cu113",
    "seaborn>=0.13.2",
    "pytest>=8.3.5",
    "numpy==1.24.3",
    "gdown==4.6.0",
    "jupyterlab>=4.3.6",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu113"
explicit = true
